+++ CHANGELOG.md
@@ -6,10 +6,12 @@ Version latest (0.12.0rc?, 2019-09-24?)
 - ...
 
 
-Version 0.12.3 (2019-10-25)
+Version 0.12.{3,4} (2019-10-25)
 
 - enh: move apiman.register.schemaValidation to apiman.request since all responses from isitfit-api will match the general schema
 - enh: add treat 'registration in progress' in the apiman.request function
+- bugfix: `isitfit tags suggest --advanced` had a bug in checking that registration was ok
+- enh: add `apiman.r_body` as shortcut
 
 
 Version 0.12.2 (2019-10-25)